+++ Rdmp.UI/AutoComplete/AutoCompleteProviderWin.cs
@@ -6,6 +6,7 @@
 
 using System;
 using System.Linq;
+using System.Windows.Forms;
 using FAnsi.Discovery.QuerySyntax;
 using Rdmp.Core.Autocomplete;
 using ScintillaNET;
@@ -21,28 +22,72 @@ namespace Rdmp.UI.AutoComplete
         {
         }
 
+        private char Separator = ';';
+
         public void RegisterForEvents(Scintilla queryEditor)
         {
+            queryEditor.KeyDown += (s, e) =>
+            {
+                if (e.KeyCode == Keys.Space && e.Control)
+                {
+                    e.SuppressKeyPress = true;
+                    ShowAutocomplete(queryEditor, true);
+                }
+            };
+            
+            queryEditor.AutoCSeparator = Separator;
             queryEditor.CharAdded += scintilla_CharAdded;
             queryEditor.AutoCIgnoreCase = true;
-        }
+            queryEditor.AutoCOrder = Order.Custom;
+            queryEditor.AutoCAutoHide = false;
 
+            for (int i = 0; i < Images.Length; i++)
+            {
+                queryEditor.RegisterRgbaImage(i, Images[i]);
+            }
+        }
 
         private void scintilla_CharAdded(object sender, CharAddedEventArgs e)
         {
             if (sender is not Scintilla scintilla)
                 return;
+            
+            ShowAutocomplete(scintilla,false);
+        }
 
+        private void ShowAutocomplete(Scintilla scintilla,bool all)
+        {
             // Find the word start
-            var currentPos = scintilla.CurrentPosition;
-            var wordStartPos = scintilla.WordStartPosition(currentPos, false);
-            var lenEntered = currentPos - wordStartPos;
-            if (lenEntered <= 0) return;
+            string word = scintilla.GetWordFromPosition(scintilla.CurrentPosition)?.Trim();
+            
+            if (string.IsNullOrWhiteSpace(word) && !all)
+            {
+                scintilla.AutoCCancel();
+                return;
+            }
 
-            var list = Items.SelectMany(GetBits).Distinct().Where(s => !string.IsNullOrWhiteSpace(s)).OrderBy(a => a);
+            var list = Items.Distinct()
+                .Where(s => !string.IsNullOrWhiteSpace(s) && s.Contains(word,StringComparison.CurrentCultureIgnoreCase))
+                .OrderBy(a => a);
+
+            if(!list.Any())
+            {
+                scintilla.AutoCCancel();
+                return;
+            }
 
             // Display the autocompletion list
-            scintilla.AutoCShow(lenEntered, string.Join(' ', list));
+            scintilla.AutoCShow(word.Length, string.Join(Separator, list.Select(FormatForAutocomplete)));
+        }
+
+        private string FormatForAutocomplete(string word)
+        {
+            if(ItemsWithImages.ContainsKey(word))
+            {
+                return $"{word}?{ItemsWithImages[word]}";
+            }
+
+            return word;
         }
     }
 }