+++ Rdmp.UI/ExtractionUIs/ExtractionInformationUI.cs
@@ -79,8 +79,6 @@ namespace Rdmp.UI.ExtractionUIs
 
             //note that we don't add the Any category
             ddExtractionCategory.DataSource = new object[] { ExtractionCategory.Core, ExtractionCategory.Supplemental, ExtractionCategory.SpecialApprovalRequired, ExtractionCategory.Internal,ExtractionCategory.Deprecated, ExtractionCategory.ProjectSpecific};
-            QueryEditor = new ScintillaTextEditorFactory().Create(new RDMPCombineableFactory());
-            QueryEditor.TextChanged += QueryEditorOnTextChanged;
 
             ObjectSaverButton1.BeforeSave += BeforeSave;
 
@@ -211,10 +209,16 @@ namespace Rdmp.UI.ExtractionUIs
                 if (ExtractionInformation.CatalogueItem.Name.Equals(ExtractionInformation.ToString()))
                     _namesMatchedWhenDialogWasLaunched = true;
 
-                var autoComplete = new AutoCompleteProviderWin(ExtractionInformation.GetQuerySyntaxHelper());
+                _querySyntaxHelper = ExtractionInformation.GetQuerySyntaxHelper();
+
+                QueryEditor = new ScintillaTextEditorFactory().Create(new RDMPCombineableFactory(), SyntaxLanguage.SQL, _querySyntaxHelper);
+                QueryEditor.TextChanged += QueryEditorOnTextChanged;
+
+                var autoComplete = new AutoCompleteProviderWin(_querySyntaxHelper);
                 autoComplete.Add(ExtractionInformation.CatalogueItem.Catalogue);
                 
                 autoComplete.RegisterForEvents(QueryEditor);
+                isFirstTimeSetupCalled = false;
             }
             
             var colInfo = ExtractionInformation.ColumnInfo;